Add ResumeNav as a class, implement the pastCPA algorithm in pastcpa.py, added scenario in scenario/ASAS/RESNAV*.scn
