+++ bluesky/traffic/asas/resolution.py
@@ -17,8 +17,6 @@ class ConflictResolution(Entity, replaceable=True):
         # [-] switch to activate priority rules for conflict resolution
         self.swprio = False  # switch priority on/off
         self.priocode = ''  # select priority mode
-        self.resopairs = set()  # Resolved conflicts that are still before CPA
-
         # Resolution factors:
         # set < 1 to maneuver only a fraction of the resolution
         # set > 1 to add a margin to separation values
@@ -43,7 +41,6 @@ class ConflictResolution(Entity, replaceable=True):
         super().reset()
         self.swprio = False
         self.priocode = ''
-        self.resopairs.clear()
         self.resofach = bs.settings.asas_marh
         self.resofacv = bs.settings.asas_marv
         self.resodhrelative = True
@@ -103,99 +100,6 @@ class ConflictResolution(Entity, replaceable=True):
             # Only perform CR when an actual method is selected
             if conf.confpairs:
                 self.trk, self.tas, self.vs, self.alt = self.resolve(conf, ownship, intruder)
-            self.resumenav(conf, ownship, intruder)
-
-    def resumenav(self, conf, ownship, intruder):
-        '''
-            Decide for each aircraft in the conflict list whether the ASAS
-            should be followed or not, based on if the aircraft pairs passed
-            their CPA.
-        '''
-        # Add new conflicts to resopairs and confpairs_all and new losses to lospairs_all
-        self.resopairs.update(conf.confpairs)
-
-        # Conflict pairs to be deleted
-        delpairs = set()
-        changeactive = dict()
-
-        # smallest relative angle between vectors of heading a and b
-        def anglediff(a, b):
-            d = a - b
-            if d > 180:
-                return anglediff(a, b + 360)
-            elif d < -180:
-                return anglediff(a + 360, b)
-            else:
-                return d
-            
-
-        # Look at all conflicts, also the ones that are solved but CPA is yet to come
-        for conflict in self.resopairs:
-            idx1, idx2 = bs.traf.id2idx(conflict)
-            # If the ownship aircraft is deleted remove its conflict from the list
-            if idx1 < 0:
-                delpairs.add(conflict)
-                continue
-
-            if idx2 >= 0:
-                # Distance vector using flat earth approximation
-                re = 6371000.
-                dist = re * np.array([np.radians(intruder.lon[idx2] - ownship.lon[idx1]) *
-                                      np.cos(0.5 * np.radians(intruder.lat[idx2] +
-                                                              ownship.lat[idx1])),
-                                      np.radians(intruder.lat[idx2] - ownship.lat[idx1])])
-
-                # Relative velocity vector
-                vrel = np.array([intruder.gseast[idx2] - ownship.gseast[idx1],
-                                 intruder.gsnorth[idx2] - ownship.gsnorth[idx1]])
-
-                # Check if conflict is past CPA
-                past_cpa = np.dot(dist, vrel) > 0.0
-
-                rpz = np.max(conf.rpz[[idx1, idx2]])
-                # hor_los:
-                # Aircraft should continue to resolve until there is no horizontal
-                # LOS. This is particularly relevant when vertical resolutions
-                # are used.
-                hdist = np.linalg.norm(dist)
-                hor_los = hdist < rpz
-
-                # Bouncing conflicts:
-                # If two aircraft are getting in and out of conflict continously,
-                # then they it is a bouncing conflict. ASAS should stay active until
-                # the bouncing stops.
-                is_bouncing = \
-                    abs(anglediff(ownship.trk[idx1], intruder.trk[idx2])) < 30.0 and \
-                    hdist < rpz * self.resofach
-
-            # Start recovery for ownship if intruder is deleted, or if past CPA
-            # and not in horizontal LOS or a bouncing conflict
-            if idx2 >= 0 and (not past_cpa or hor_los or is_bouncing):
-                # Enable ASAS for this aircraft
-                changeactive[idx1] = True
-            else:
-                # Switch ASAS off for ownship if there are no other conflicts
-                # that this aircraft is involved in.
-                changeactive[idx1] = changeactive.get(idx1, False)
-                # If conflict is solved, remove it from the resopairs list
-                delpairs.add(conflict)
-
-        for idx, active in changeactive.items():
-            # Loop a second time: this is to avoid that ASAS resolution is
-            # turned off for an aircraft that is involved simultaneously in
-            # multiple conflicts, where the first, but not all conflicts are
-            # resolved.
-            self.active[idx] = active
-            if not active:
-                # Waypoint recovery after conflict: Find the next active waypoint
-                # and send the aircraft to that waypoint.
-                iwpid = bs.traf.ap.route[idx].findact(idx)
-                if iwpid != -1:  # To avoid problems if there are no waypoints
-                    bs.traf.ap.route[idx].direct(
-                        idx, bs.traf.ap.route[idx].wpname[iwpid])
-
-        # Remove pairs from the list that are past CPA or have deleted aircraft
-        self.resopairs -= delpairs
 
     @command(name='PRIORULES')
     def setprio(self, flag : bool = None, priocode=''):
